+++ frontend/components/Landing/landing.less
@@ -1,0 +1,23 @@
+.mainDiv {
+    background: lightblue;
+    position: fixed;
+    top: 0;
+    left: 0;
+    min-width: 100%;
+    min-height: 100%;
+        .header {
+            display: flex;
+            margin: 2rem 3rem 0rem 3rem;
+            justify-content: space-between;
+        }
+        .mainContent {
+           position: fixed;
+           top: 40%;
+           left: 30%;
+           right: 30%;
+           font-weight: bold;
+           font-size: 3.4rem;
+           line-height: 3.5rem;
+           text-align: center;
+        }
+}
